# openapi.yaml
...
openapi: 3.1.0
info:
  title: Kryton API
  version: 1.1.0
  description: |
    Kryton Windows virtualization control plane. All machine operations are project-scoped.
    Authenticate with `Authorization: Bearer <token>` (apikey mode) or trusted proxy headers.
    OpenAPI is served at `/openapi.yaml` and discovery at `GET /api/v1`.
  contact:
    name: Zyvor
    url: https://zyvor.dev
  license:
    name: Apache-2.0
servers:
  - url: /
    description: Relative to Kryton base URL
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/doctor:
    get:
      tags: [meta]
      summary: Environment diagnostics
      responses:
        '200': {description: Healthy}
        '503': {description: One or more checks failed}
  /api/v1/settings:
    get:
      tags: [settings]
      summary: Runtime and system settings
      responses:
        '200': {description: Settings view}
    put:
      tags: [settings]
      summary: Update runtime settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsUpdate'
      responses:
        '200': {description: Updated settings}
        '400': {description: Invalid request}
        '403': {description: Forbidden}
  /api/v1/settings/test:
    post:
      tags: [settings]
      summary: Test control-plane connectivity
      responses:
        '200': {description: All probes passed}
        '503': {description: One or more probes failed}
  /api/v1/integrations/atlas/test:
    post:
      tags: [integrations]
      summary: Test Atlas storage control plane integration
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtlasConfig'
      responses:
        '200': {description: Atlas reachable}
        '503': {description: Atlas probes failed}
...
  /api/v1/storage/config:
    get:
      tags: [storage]
      summary: Active default StorageClass
      responses:
        '200': {description: Storage config}
    put:
      tags: [storage]
      summary: Set default StorageClass for new VM disks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                storageClass: {type: string, example: rook-ceph-block}
      responses:
        '200': {description: Updated}
        '400': {description: Invalid class}
  /api/v1/storage/setup:
    get:
      tags: [storage]
      summary: Storage install job status and logs
      responses:
        '200': {description: Setup status}
    post:
      tags: [storage]
      summary: Install Longhorn or Rook Ceph
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StorageSetup'
      responses:
        '202': {description: Setup started}
        '400': {description: Invalid request}
...
  /api/v1/golden:
    get:
      tags: [images]
      summary: Golden image builds
      responses:
        '200': {description: Builds}
    post:
      tags: [images]
      summary: Start golden image build
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageId: {type: string}
                version: {type: string}
                auto: {type: boolean}
      responses:
        '202': {description: Build started}
  /api/v1/golden/{id}:
    get:
      tags: [images]
      parameters:
        - $ref: '#/components/parameters/GoldenID'
      responses:
        '200': {description: Build}
        '404': {description: Not found}
  /api/v1/golden/{id}/passport:
    get:
      tags: [images]
      summary: guestkit Cutover Passport (boot/migration evidence) for this build
      parameters:
        - $ref: '#/components/parameters/GoldenID'
      responses:
        '200': {description: Passport JSON}
        '404': {description: No passport recorded for this build}
  /api/v1/golden/{id}/bootstrap:
    post:
      tags: [images]
      parameters:
        - $ref: '#/components/parameters/GoldenID'
      responses:
        '202': {description: CDI bootstrap started}
        '400': {description: Build not ready}
...
  /api/v1/jobs/{id}:
    get:
      tags: [jobs]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        '200': {description: Job}
        '404': {description: Not found}
  /api/v1/summary:
    get:
      tags: [machines]
      parameters:
        - $ref: '#/components/parameters/Project'
      responses:
        '200': {description: Project summary}
  /api/v1/machines:
    get:
      tags: [machines]
      parameters:
        - $ref: '#/components/parameters/Project'
        - in: query
          name: limit
          description: Max items per page (1-500).
          schema: {type: integer, default: 50}
        - in: query
          name: cursor
          description: Opaque page cursor from a previous response's nextCursor.
          schema: {type: string}
      responses:
        '200': {description: Machines}
    post:
      tags: [machines]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMachine'
      responses:
        '201': {description: Machine created}
        '400': {description: Invalid request}
        '409': {description: Name conflict}
  /api/v1/machines/{id}:
    parameters:
      - $ref: '#/components/parameters/MachineID'
      - $ref: '#/components/parameters/Project'
    get:
      tags: [machines]
      responses:
        '200': {description: Machine}
        '404': {description: Not found}
    delete:
      tags: [machines]
      responses:
        '204': {description: Deleted}
  /api/v1/machines/{id}/start:
    post:
      tags: [machines]
      parameters:
        - $ref: '#/components/parameters/MachineID'
        - $ref: '#/components/parameters/Project'
      responses:
        '200': {description: Start requested}
  /api/v1/machines/{id}/stop:
    post:
      tags: [machines]
      parameters:
        - $ref: '#/components/parameters/MachineID'
        - $ref: '#/components/parameters/Project'
      responses:
        '200': {description: Stop requested}
  /api/v1/machines/{id}/snapshot:
    post:
      tags: [machines]
      parameters:
        - $ref: '#/components/parameters/MachineID'
        - $ref: '#/components/parameters/Project'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: {type: string}
      responses:
        '201': {description: Snapshot requested}
  /api/v1/machines/{id}/snapshots:
    get:
      tags: [machines]
      parameters:
        - $ref: '#/components/parameters/MachineID'
        - $ref: '#/components/parameters/Project'
      responses:
        '200': {description: Snapshot list}
  /api/v1/machines/{id}/snapshots/{sid}/restore:
    post:
      tags: [machines]
      parameters:
        - $ref: '#/components/parameters/MachineID'
        - $ref: '#/components/parameters/SnapshotID'
        - $ref: '#/components/parameters/Project'
      responses:
        '202': {description: Restore requested}
  /api/v1/machines/{id}/snapshots/{sid}:
    delete:
      tags: [machines]
      parameters:
        - $ref: '#/components/parameters/MachineID'
        - $ref: '#/components/parameters/SnapshotID'
        - $ref: '#/components/parameters/Project'
      responses:
        '204': {description: Snapshot deleted}
  /api/v1/machines/{id}/console:
    get:
      tags: [machines]
      summary: Web console (provider-dependent)
      parameters:
        - $ref: '#/components/parameters/MachineID'
        - $ref: '#/components/parameters/Project'
      responses:
        '200': {description: Console HTML or redirect}
  /api/v1/machines/{id}/vnc:
    get:
      tags: [machines]
      summary: VNC websocket/proxy (KubeVirt)
      parameters:
        - $ref: '#/components/parameters/MachineID'
        - $ref: '#/components/parameters/Project'
      responses:
        '200': {description: VNC stream}
  /api/v1/events:
    get:
      tags: [events]
      parameters:
        - in: query
          name: limit
          schema: {type: integer, default: 100}
      responses:
        '200': {description: CloudEvent history}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key from KRYTON_API_KEYS_FILE (hash with krytonctl hash-token)
  parameters:
    Project:
      in: query
      name: project
      schema: {type: string}
      description: Project scope (defaults to identity default when omitted on some routes)
    MachineID:
      in: path
      name: id
      required: true
      schema: {type: string, format: uuid}
    SnapshotID:
      in: path
      name: sid
      required: true
      schema: {type: string}
    GoldenID:
      in: path
      name: id
      required: true
      schema: {type: string}
  schemas:
    CreateMachine:
      type: object
      required: [project, name, image, compute, disk]
      properties:
        project: {type: string}
        name: {type: string}
        image: {type: string}
        compute:
          type: object
          required: [cpu, memoryMiB]
          properties:
            cpu: {type: integer, minimum: 1, maximum: 256}
            memoryMiB: {type: integer, minimum: 512}
        disk:
          type: object
          required: [sizeGiB]
          properties:
            sizeGiB: {type: integer, minimum: 16}
            storageClass: {type: string, description: Override Settings default (e.g. rook-ceph-block)}
        network:
          type: object
          properties:
            networkId: {type: string}
        ttlMinutes: {type: integer, minimum: 0, maximum: 43200}
        initialization:
          $ref: "#/components/schemas/LinuxInitialization"
        dockur:
          $ref: '#/components/schemas/DockurOptions'
    LinuxInitialization:
      type: object
      additionalProperties: false
      properties:
        username: {type: string, pattern: "^[a-z_][a-z0-9_-]{0,31}$"}
        sshAuthorizedKeys:
          type: array
          maxItems: 32
          items: {type: string, maxLength: 16384}
...
    SettingsUpdate:
      type: object
      properties:
        defaultProject: {type: string}
        imageNamespace: {type: string}
        eventWebhookUrl: {type: string}
        storageClass: {type: string}
    StorageSetup:
      type: object
      required: [backend]
      properties:
        backend: {type: string, enum: [longhorn, rook-ceph]}
        rookMode: {type: string, enum: [lab, devices, pool-only, device]}
        device: {type: string, example: /dev/sdb1}
        wipeDevice: {type: boolean}
        setDefault: {type: boolean, default: true}
    AtlasConfig:
      type: object
      properties:
        enabled: {type: boolean}
        baseUrl: {type: string, example: http://127.0.0.1:5110}
        token: {type: string}
        product: {type: string, default: kryton}
        preferAtlas: {type: boolean}
...